Big update
Here's basically the entire project. Everything seems to be working to
some degree at least. I'll try to add encryption tomorrow morning if I
have time before class.

It doesn't use a router which is something he may want us to change
about it.

It also doesn't save or load the .card files. It just has a few hard
coded in for now. I'm not sure if he'll notice or care about that or
not.

There may be some small bugs, but I think the basic functionality is all
here.